requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
maxminddb==2.0.3
maxminddb-geolite2==2018.703
mysqlclient==2.0.1
Pillow==8.1.0
python-geoip==1.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
